Slescs/backend/auth_repo/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.98
botocore==1.34.98
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
crayons==0.4.0
cryptography==42.0.5
defusedxml==0.7.1
disutils==1.4.32.post2
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==3.2.1
django-allauth==0.61.1
django-cors-headers==4.3.1
django-storages==1.14.3
django-environ==0.11.2
django-rest-passwordreset==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
grpcio==1.62.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
IMAPClient==3.0.1
jmespath==1.0.1
jsonrpclib==0.2.1
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
piexif==1.1.3
pillow==10.2.0
plumbum==1.8.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rpyc==6.0.0
s3transfer==0.10.1
selenium==4.19.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stegano==0.11.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.4
